Make dispatchEvent's event object currentTime property change-tracking
* make pluginMediaElement.currentTime a get-only property outside of
timeupdate's, so that we are guaranteed to be comparing the latest value
from the player with the last polled value. I believe this was the
original intention of the property anyway based on the comment.

* decorate the value of obj.currentTime with a 'changed' property, so
that callers can ignore 'timeupdates' where the event fires, but the
currentTime of the player hasn't changed.